Alzheimers_Project/MRI/requirements.txt
absl-py==0.9.0
alabaster==0.7.12
altair==4.2.0
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
asn1crypto==0.24.0
astor==0.8.1
astroid==2.2.5
astropy==3.2.1
astunparse==1.6.3
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
backports.zoneinfo==0.2.1
base58==2.1.1
beautifulsoup4==4.7.1
bitarray==0.9.3
bkcharts==0.2
bleach==3.1.0
blinker==1.4
bokeh==1.2.0
boto==2.49.0
Bottleneck==1.2.1
cachetools==5.0.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
clyent==1.2.2
colorama==0.4.1
comtypes==1.1.7
conda==4.7.10
conda-build==3.18.8
conda-package-handling==1.3.11
conda-verify==3.4.2
contextlib2==0.5.5
cryptography==2.7
cycler==0.10.0
Cython==0.29.12
cytoolz==0.10.0
dask==2.1.0
debugpy==1.5.1
decorator==4.4.0
defusedxml==0.6.0
distributed==2.1.0
docopt==0.6.2
docutils==0.14
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
Flask==1.1.1
flatbuffers==2.0
fonttools==4.31.1
future==0.17.1
gast==0.5.3
gevent==1.4.0
gitdb==4.0.9
GitPython==3.1.27
glob2==0.7
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.1.8
greenlet==0.4.15
grpcio==1.26.0
h5py==2.9.0
heapdict==1.0.0
html5lib==1.0.1
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==4.11.2
inflection==0.3.1
ipykernel==6.9.1
ipython==7.32.0
ipython_genutils==0.2.0
ipywidgets==7.5.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.18.1
Jinja2==2.10.1
joblib==0.13.2
json5==0.8.4
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.0.2
jupyterlab-server==1.0.0
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
libarchive-c==2.8
libclang==13.0.0
llvmlite==0.29.0
locket==0.2.0
lxml==4.3.4
Markdown==3.3.6
MarkupSafe==1.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
mkl-service==2.0.2
mock==3.0.5
more-itertools==7.0.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.5.0
nbformat==4.4.0
nest-asyncio==1.5.4
networkx==2.3
nltk==3.4.4
nose==1.3.7
notebook==6.0.0
numba==0.44.1
numexpr==2.6.9
numpy==1.21.5
numpydoc==0.9.1
oauthlib==3.2.0
olefile==0.46
opencv-python==4.5.5.64
openpyxl==2.6.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.4.2
parso==0.8.3
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.4
patsy==0.5.1
pep8==1.7.1
pickleshare==0.7.5
Pillow==9.0.1
pipreqs==0.4.11
pkginfo==1.5.0.1
plotly==5.6.0
pluggy==0.12.0
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.19.4
psutil==5.6.3
py==1.8.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
pydeck==0.7.1
pyflakes==2.1.1
pygame==2.1.2
Pygments==2.4.2
pylint==2.3.1
Pympler==1.0.1
pyodbc==4.0.26
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyreadline==2.1
pyrsistent==0.14.11
PySocks==1.7.0
pytest==5.0.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.8.0
pytz==2019.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.0.3
pywin32==223
pywinpty==0.5.5
PyYAML==5.1.1
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.5.1
QtPy==1.8.0
Quandl==3.5.0
requests==2.22.0
requests-oauthlib==1.3.1
rope==0.14.0
rsa==4.8
ruamel_yaml==0.15.46
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.7.3
seaborn==0.9.0
semver==2.13.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.13.0
sklearn==0.0
smmap==5.0.0
snowballstemmer==1.9.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.1
SQLAlchemy==1.3.5
statsmodels==0.10.0
streamlit==1.7.0
sympy==1.4
tables==3.5.2
tblib==1.4.0
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf-estimator-nightly==2.8.0.dev2021122109
toml==0.10.2
toolz==0.10.0
tornado==6.0.3
tqdm==4.32.1
traitlets==4.3.2
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
unicodecsv==0.14.1
urllib3==1.24.2
validators==0.18.2
watchdog==2.1.6
wavio==0.0.4
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.5.0
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.1.8
xlwings==0.15.8
xlwt==1.3.0
yarg==0.1.9
zict==1.0.0
zipp==0.5.1
